Make secret storage pluggable
* Create Secret and SecretStore Protocol classes
* Create AgeSecret and AgeSecretStore implementation classes
* Separate inventory from secret store
* Remove sitewrapper.set_site_module_filepath(). See reasoning below.
* Add some debug logging

Why remove sitewrapper.set_site_module_filepath()?

It relied on importing the site module, but with this commit, the site
module's root file has to import progfiguration core stuff like
Inventory and SecretStore. (We will provide a useful inventory and
Age-based secret store in the core package, even as we allow sites to
define their own.) This meant that progfiguration core got imported
before we knew the file path for the site package, which meant that
progfiguration.progfigsite_module_path was always the default value
"progfigsite".

The only place you'd need to call this is in a site's entry point. The
site is designed to need a single entrypoint script (the shim), which
didn't have this issue because it modified sys.path. Other entry points
defined in pyproject.toml are only needed for development mode, and will
continue to work. The only other place we need this is in tests, so we
just use the same sys.path manipulation in the test setup and it works.
Hence, no need for set_site_module_filepath().