Implement progfigsite.version.get_version()
* Remove requirement of a get_version() in site root modules
* Add requirement of a version submodule with get_version() in it
* Remove get_version_factory_with_simple_fallback()

The driver for this change is the ability to
`pip install -e '.[development]'`.
The 'development' extras are expected to include progfiguration core,
but before installing them 'pip' first calculates the site's version.
If this is stored in the site's root module, which must reference
progfiguration core's sitewrapper, and may reference inventory helpers,
pip will fail. Instead, we now require a separate module that pip will
import directly that can calculate the version.